changed ints to doubles, do to fact that the average grade might end up being a double. Same done in unit tests

+++ Onsdagsopgave6/src/Student.java
@@ -4,9 +4,9 @@ public class Student {
 
     private String name;
     private int graduateYear;
-    private int [] grades;
+    private double [] grades;
 
-    public Student(String name, int graduateYear, int grades [] ){
+    public Student(String name, int graduateYear, double grades [] ){
 
         this.name = name;
         this.graduateYear = graduateYear;
@@ -22,9 +22,9 @@ public class Student {
                 "\nGrades= " + Arrays.toString(grades);
     }
 
-    public int averageGrade(){
+    public double averageGrade(){
 
-        int sum = (grades[0] + grades[1] +grades[2] + grades[3] +grades[4] + grades[5] + grades[6]) / 7;
+        double sum = (grades[0] + grades[1] +grades[2] + grades[3] +grades[4] + grades[5] + grades[6]) / 7;
         System.out.println("Average grade is: " + sum + "\n");
         return sum;
     }
